refactor: 패키지명 popup.info로 수정
