+++ info/src/main/java/com/popup/info/domain/PopUpStoreInfo.java
@@ -1,4 +1,4 @@
-package com.example.InfoCrawler.domain;
+package com.popup.info.domain;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
