+++ info/src/main/java/com/popup/info/repository/InfoRepository.java
@@ -1,10 +1,10 @@
-package com.example.InfoCrawler.repository;
+package com.popup.info.repository;
 
 
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
-import com.example.InfoCrawler.domain.PopUpStoreInfo;
+import com.popup.info.domain.PopUpStoreInfo;
 
 @Repository
 public interface InfoRepository extends MongoRepository<PopUpStoreInfo, String> {
