+++ info/src/main/java/com/popup/info/service/WebDriverManager.java
@@ -1,4 +1,4 @@
-package com.example.InfoCrawler.service;
+package com.popup.info.service;
 
 import java.io.File;
